Update Swagger support to 2.0
swagger-codegen throws NullPointer exceptions when trying to parse the
Swagger 1.x output generated by loopback-explorer. This is probably
caused by Loopback generating an invalid swagger specification. It
refers to the invalid type “ref”.

To get Swagger 2.0 support we upgrade to loopback-component-explorer
2.3.0.

This requires a bit of updating in how it is mounted as well.